Adding bootstrap and improving feedback
Feedback now pops up a bootstrap modal to submit. It also correctly makes the
feedback file if it doesn't exist.

It also adds a way for the frontend to send things to the backend that aren't
user-typed things and things that don't need a response.

This also correctly disables clicking on previous equations in all cases.

This also adds a small border around the page.