Restructured tagging
Now the boxes are drawn whenever a tag is seen in the textbox.

There theoretically may be performance issues because all of them are redrawn
for every character typed, but I haven't noticed anything bad, and lines should
be pretty short anyway.

Click and drag to select parent is implemented, but siblings are not yet.

Colors change when tags are inserted early, which I'm pretty sure is solvable.