Adding download/load session buttons
Can save and restore sessions this way. Also fixed a bunch of concurrency bugs
that became apparent when things are going more quickly than mathjax can
respond. There still exists a bug (if the server sends non-math before the
previous math is finished rendering, the math corresponding to the non-math
output is not ever rendered), but I will solve that by making the server always
send math in a future commit.