Running cargo fmt
This touches a lot of stuff, so if we want to revert or rebase commits past
here, we should just delete this commit, too.

Note that cargo fmt fails on the pub_restricted syntax (i.e. pub(super), so in
the future when running cargo fmt, we need to manually fix this